lint and test fixes

+++ tomplot/tomplot_tools.py
@@ -370,7 +370,7 @@ def roundup(number, digits=0):
     """
 
     n = 10**-digits
-    return round(np.ceil(number / n) * n, digits)
+    return round(np.ceil(number / n) * n, int(digits))
 
 
 def rounddown(number, digits=0):
@@ -386,7 +386,7 @@ def rounddown(number, digits=0):
     """
 
     n = 10**-digits
-    return round(np.floor(number / n) * n, digits)
+    return round(np.floor(number / n) * n, int(digits))
 
 
 def tomplot_contours(data, min_num_bins=10, divergent_flag=False,
@@ -435,7 +435,7 @@ def tomplot_contours(data, min_num_bins=10, divergent_flag=False,
         while discrete_diff > 2.0*(raw_max - raw_min):
 
             if (raw_max - raw_min > 0):
-                digits = np.floor(-np.log10(raw_max - raw_min)) + digit_inc
+                digits = int(np.floor(-np.log10(raw_max - raw_min)) + digit_inc)
             else:
                 digits = 1
 
@@ -459,9 +459,9 @@ def tomplot_contours(data, min_num_bins=10, divergent_flag=False,
     else:
         # Data is considered to be zero
         if raw_max > 0:
-            digits = np.floor(-np.log10(raw_max)) + 3
+            digits = int(np.floor(-np.log10(raw_max)) + 3)
         elif raw_min < 0:
-            digits = np.floor(-np.log10(- raw_min)) + 3
+            digits = int(np.floor(-np.log10(- raw_min)) + 3)
         else:
             digits = 3
 
@@ -486,7 +486,7 @@ def tomplot_contours(data, min_num_bins=10, divergent_flag=False,
 
     # To be safe, only make contours when we have non-zero data
     if (raw_max - raw_min) > zero_tolerance:
-        step_digits = np.floor(-np.log10(max_step))
+        step_digits = int(np.floor(-np.log10(max_step)))
         # Find an even bigger maximum step by rounding max_step up
         max_max_step = roundup(max_step, digits=step_digits)
 
@@ -803,4 +803,3 @@ def tomplot_legend_ax(ax, location='top', extra_padding=0.0, **leg_kwargs):
 
     # Set new position
     leg.set_bbox_to_anchor(new_bbox)
-
